+++ src/publisher/tf_broadcaster.cpp
@@ -13,8 +13,8 @@ TFBroadCaster::TFBroadCaster(std::string frame_id, std::string child_frame_id) {
   transform_.child_frame_id_ = child_frame_id;
 }
 
-void TFBroadCaster::sendTransform(Eigen::Matrix4f pose, double time) {
-  Eigen::Quaternionf q(pose.block<3, 3>(0, 0));
+void TFBroadCaster::sendTransform(Eigen::Matrix4d pose, double time) {
+  Eigen::Quaterniond q(pose.block<3, 3>(0, 0));
   ros::Time ros_time(static_cast<float>(time));
   transform_.stamp_ = ros_time;
   transform_.setRotation(tf::Quaternion(q.x(), q.y(), q.z(), q.w()));
